+++ src/CardShopCoop/CoopPlugin.cs
@@ -12,7 +12,7 @@ namespace CardShopCoop
     {
         public const string Guid = "com.zwhit.cardshopcoop";
         public const string Name = "CardShopCoop";
-        public const string Version = "1.0.39";
+        public const string Version = "1.0.40";
 
         public static ManualLogSource Log;